refactor(coding-agent): split daemon mode modules
