+++ packages/coding-agent/src/modes/daemon/daemon-mode.ts
@@ -7,209 +7,40 @@
  */
 
 import { randomUUID } from "node:crypto";
-import { existsSync, lstatSync, unlinkSync } from "node:fs";
-import { createConnection, createServer, type Server, type Socket } from "node:net";
-import { tmpdir } from "node:os";
-import { join, resolve } from "node:path";
-import type { AgentMessage, ThinkingLevel } from "@earendil-works/pi-agent-core";
-import type { Api, ImageContent, Model } from "@earendil-works/pi-ai";
-import type { AgentSession, AgentSessionEvent, SessionStats } from "../../core/agent-session.js";
-import {
-	type AgentSessionRuntime,
-	type CreateAgentSessionRuntimeFactory,
-	createAgentSessionRuntime,
-} from "../../core/agent-session-runtime.js";
-import type {
-	ExtensionCommandContextActions,
-	ExtensionUIContext,
-	ExtensionUIDialogOptions,
-	ExtensionWidgetOptions,
-	WorkingIndicatorOptions,
-} from "../../core/extensions/index.js";
+import { createServer, type Server, type Socket } from "node:net";
+import { resolve } from "node:path";
+import type { SessionStats } from "../../core/agent-session.js";
+import { type AgentSessionRuntime, createAgentSessionRuntime } from "../../core/agent-session-runtime.js";
 import { type SessionInfo, SessionManager } from "../../core/session-manager.js";
 import { killTrackedDetachedChildren } from "../../utils/shell.js";
-import { type Theme, theme } from "../interactive/theme/theme.js";
 import { attachJsonlLineReader, serializeJsonLine } from "../rpc/jsonl.js";
 import type { RpcSlashCommand } from "../rpc/rpc-types.js";
+import { bindDaemonSessionRecord } from "./daemon-extension-binding.js";
+import {
+	type DaemonCommand,
+	type DaemonModeOptions,
+	type DaemonOutbound,
+	type DaemonResponse,
+	failure,
+	success,
+} from "./daemon-protocol.js";
+import {
+	type DaemonSessionRecord,
+	type DaemonSocketClient,
+	stateForSession,
+	summaryForRecord,
+} from "./daemon-session-record.js";
+import { cleanupDaemonSocketPath, defaultDaemonSocketPath, prepareDaemonSocketPath } from "./daemon-socket.js";
 
-export interface DaemonModeOptions {
-	socketPath?: string;
-	cwd: string;
-	agentDir: string;
-	sessionDir?: string;
-	createRuntime: CreateAgentSessionRuntimeFactory;
-}
-
-export interface DaemonSessionSummary {
-	activeSessionId: string;
-	sessionId: string;
-	sessionFile?: string;
-	sessionName?: string;
-	cwd: string;
-	isStreaming: boolean;
-	isCompacting: boolean;
-	attachedClients: number;
-	messageCount: number;
-}
-
-export interface DaemonSessionState {
-	model?: Model<Api>;
-	thinkingLevel: ThinkingLevel;
-	isStreaming: boolean;
-	isCompacting: boolean;
-	sessionFile?: string;
-	sessionId: string;
-	sessionName?: string;
-	messageCount: number;
-	pendingMessageCount: number;
-	streamingMessage?: AgentMessage;
-}
-
-export type DaemonCommand =
-	| { id?: string; type: "list" }
-	| { id?: string; type: "list_saved"; cwd?: string; sessionDir?: string }
-	| {
-			id?: string;
-			type: "create";
-			cwd?: string;
-			sessionDir?: string;
-			sessionPath?: string;
-			continueRecent?: boolean;
-			name?: string;
-	  }
-	| { id?: string; type: "attach"; activeSessionId: string }
-	| { id?: string; type: "detach"; activeSessionId?: string }
-	| { id?: string; type: "kill"; activeSessionId: string }
-	| { id?: string; type: "rename"; activeSessionId: string; name: string }
-	| {
-			id?: string;
-			type: "prompt";
-			activeSessionId: string;
-			message: string;
-			images?: ImageContent[];
-			streamingBehavior?: "steer" | "followUp";
-	  }
-	| { id?: string; type: "steer"; activeSessionId: string; message: string; images?: ImageContent[] }
-	| { id?: string; type: "follow_up"; activeSessionId: string; message: string; images?: ImageContent[] }
-	| { id?: string; type: "abort"; activeSessionId: string }
-	| { id?: string; type: "get_state"; activeSessionId: string }
-	| { id?: string; type: "get_messages"; activeSessionId: string }
-	| { id?: string; type: "get_session_stats"; activeSessionId: string }
-	| { id?: string; type: "get_commands"; activeSessionId: string }
-	| { id?: string; type: "shutdown" };
-
-type DaemonCommandName = DaemonCommand["type"];
-
-export type DaemonResponse =
-	| { id?: string; type: "response"; command: string; success: true; data?: unknown }
-	| { id?: string; type: "response"; command: string; success: false; error: string };
-
-export type DaemonOutbound =
-	| DaemonResponse
-	| { type: "daemon_hello"; socketPath: string }
-	| { type: "session_event"; activeSessionId: string; event: AgentSessionEvent }
-	| { type: "session_attached"; activeSessionId: string; state: DaemonSessionState; messages: AgentMessage[] }
-	| { type: "session_detached"; activeSessionId: string }
-	| { type: "session_closed"; activeSessionId: string; reason: "killed" | "shutdown" }
-	| {
-			type: "extension_ui_request";
-			activeSessionId: string;
-			method: string;
-			payload: Record<string, unknown>;
-	  }
-	| { type: "extension_error"; activeSessionId: string; extensionPath: string; event: string; error: string };
-
-interface DaemonClient {
-	id: string;
-	socket: Socket;
-	attachedActiveSessionIds: Set<string>;
-	detachInput: () => void;
-}
-
-interface DaemonSessionRecord {
-	activeSessionId: string;
-	runtime: AgentSessionRuntime;
-	clients: Set<DaemonClient>;
-	unsubscribe?: () => void;
-}
-
-export function defaultDaemonSocketPath(): string {
-	if (process.platform === "win32") {
-		return `\\\\.\\pipe\\prime-agent-${process.pid}`;
-	}
-	return join(tmpdir(), "prime-agent-daemon.sock");
-}
-
-function stateForSession(session: AgentSession): DaemonSessionState {
-	return {
-		model: session.model as Model<Api> | undefined,
-		thinkingLevel: session.thinkingLevel,
-		isStreaming: session.isStreaming,
-		isCompacting: session.isCompacting,
-		sessionFile: session.sessionFile,
-		sessionId: session.sessionId,
-		sessionName: session.sessionName,
-		messageCount: session.messages.length,
-		pendingMessageCount: session.pendingMessageCount,
-		streamingMessage: session.state.streamingMessage,
-	};
-}
-
-function summaryForRecord(record: DaemonSessionRecord): DaemonSessionSummary {
-	const session = record.runtime.session;
-	return {
-		activeSessionId: record.activeSessionId,
-		sessionId: session.sessionId,
-		sessionFile: session.sessionFile,
-		sessionName: session.sessionName,
-		cwd: session.sessionManager.getCwd(),
-		isStreaming: session.isStreaming,
-		isCompacting: session.isCompacting,
-		attachedClients: record.clients.size,
-		messageCount: session.messages.length,
-	};
-}
-
-function success(id: string | undefined, command: DaemonCommandName, data?: unknown): DaemonResponse {
-	return data === undefined
-		? { id, type: "response", command, success: true }
-		: { id, type: "response", command, success: true, data };
-}
-
-function failure(id: string | undefined, command: string, error: unknown): DaemonResponse {
-	return {
-		id,
-		type: "response",
-		command,
-		success: false,
-		error: error instanceof Error ? error.message : String(error),
-	};
-}
-
-function canConnectToUnixSocket(socketPath: string): Promise<boolean> {
-	return new Promise((resolveConnect) => {
-		const socket = createConnection(socketPath);
-		let settled = false;
-		let timeoutId: ReturnType<typeof setTimeout> | undefined;
-
-		const finish = (canConnect: boolean) => {
-			if (settled) {
-				return;
-			}
-			settled = true;
-			if (timeoutId) {
-				clearTimeout(timeoutId);
-			}
-			socket.removeAllListeners();
-			socket.destroy();
-			resolveConnect(canConnect);
-		};
-
-		timeoutId = setTimeout(() => finish(false), 250);
-		socket.once("connect", () => finish(true));
-		socket.once("error", () => finish(false));
-	});
-}
+export type {
+	DaemonCommand,
+	DaemonModeOptions,
+	DaemonOutbound,
+	DaemonResponse,
+	DaemonSessionState,
+	DaemonSessionSummary,
+} from "./daemon-protocol.js";
+export { defaultDaemonSocketPath } from "./daemon-socket.js";
 
 export async function runDaemonMode(initialRuntime: AgentSessionRuntime, options: DaemonModeOptions): Promise<never> {
 	const socketPath = options.socketPath ?? defaultDaemonSocketPath();
@@ -226,7 +57,7 @@ class AgentDaemon {
 	private server?: Server;
 	private shuttingDown = false;
 	private ownsSocketPath = false;
-	private readonly clients = new Set<DaemonClient>();
+	private readonly clients = new Set<DaemonSocketClient>();
 	private readonly sessions = new Map<string, DaemonSessionRecord>();
 	private readonly signalCleanupHandlers: Array<() => void> = [];
 
@@ -236,7 +67,7 @@ class AgentDaemon {
 	) {}
 
 	async start(): Promise<void> {
-		await this.prepareSocketPath();
+		await prepareDaemonSocketPath(this.socketPath);
 
 		this.server = createServer((socket) => this.handleConnection(socket));
 
@@ -266,35 +97,12 @@ class AgentDaemon {
 		console.error(`Prime Agent daemon listening on ${this.socketPath}`);
 	}
 
-	private async prepareSocketPath(): Promise<void> {
-		if (process.platform === "win32" || !existsSync(this.socketPath)) {
-			return;
-		}
-
-		const stat = lstatSync(this.socketPath);
-		if (!stat.isSocket()) {
-			throw new Error(`Daemon socket path exists and is not a socket: ${this.socketPath}`);
-		}
-
-		if (await canConnectToUnixSocket(this.socketPath)) {
-			throw new Error(`Daemon socket already in use: ${this.socketPath}`);
-		}
-
-		unlinkSync(this.socketPath);
-	}
-
 	private cleanupSocketPath(): void {
-		if (process.platform === "win32" || !this.ownsSocketPath) {
+		if (!this.ownsSocketPath) {
 			return;
 		}
 		this.ownsSocketPath = false;
-		try {
-			if (existsSync(this.socketPath)) {
-				unlinkSync(this.socketPath);
-			}
-		} catch {
-			// Best effort cleanup; shutdown should not be blocked by socket unlink failures.
-		}
+		cleanupDaemonSocketPath(this.socketPath);
 	}
 
 	private async addRuntime(runtime: AgentSessionRuntime, name?: string): Promise<DaemonSessionRecord> {
@@ -303,7 +111,12 @@ class AgentDaemon {
 			runtime,
 			clients: new Set(),
 		};
-		await this.bindRecord(record);
+		await bindDaemonSessionRecord(record, {
+			broadcast: (targetRecord, message) => this.broadcastToRecord(targetRecord, message),
+			shutdown: () => {
+				void this.shutdown(0);
+			},
+		});
 		this.sessions.set(record.activeSessionId, record);
 		if (name) {
 			record.runtime.session.setSessionName(name);
@@ -327,153 +140,6 @@ class AgentDaemon {
 		return this.addRuntime(runtime, command.name);
 	}
 
-	private async bindRecord(record: DaemonSessionRecord): Promise<void> {
-		const session = record.runtime.session;
-
-		record.unsubscribe?.();
-		record.unsubscribe = session.subscribe((event) => {
-			this.broadcastToRecord(record, {
-				type: "session_event",
-				activeSessionId: record.activeSessionId,
-				event,
-			});
-		});
-
-		record.runtime.setRebindSession(async () => {
-			await this.bindRecord(record);
-		});
-
-		await session.bindExtensions({
-			uiContext: this.createExtensionUIContext(record),
-			commandContextActions: this.createCommandContextActions(record),
-			shutdownHandler: () => {
-				void this.shutdown(0);
-			},
-			onError: (error) => {
-				this.broadcastToRecord(record, {
-					type: "extension_error",
-					activeSessionId: record.activeSessionId,
-					extensionPath: error.extensionPath,
-					event: error.event,
-					error: error.error,
-				});
-			},
-		});
-	}
-
-	private createCommandContextActions(record: DaemonSessionRecord): ExtensionCommandContextActions {
-		return {
-			waitForIdle: () => record.runtime.session.agent.waitForIdle(),
-			newSession: async (options) => record.runtime.newSession(options),
-			fork: async (entryId, options) => {
-				const result = await record.runtime.fork(entryId, options);
-				return { cancelled: result.cancelled };
-			},
-			navigateTree: async (targetId, options) => {
-				const result = await record.runtime.session.navigateTree(targetId, {
-					summarize: options?.summarize,
-					customInstructions: options?.customInstructions,
-					replaceInstructions: options?.replaceInstructions,
-					label: options?.label,
-				});
-				return { cancelled: result.cancelled };
-			},
-			switchSession: async (sessionPath, options) => record.runtime.switchSession(sessionPath, options),
-			reload: async () => {
-				await record.runtime.session.reload();
-			},
-		};
-	}
-
-	private createExtensionUIContext(record: DaemonSessionRecord): ExtensionUIContext {
-		const emitUiRequest = (method: string, payload: Record<string, unknown>): void => {
-			this.broadcastToRecord(record, {
-				type: "extension_ui_request",
-				activeSessionId: record.activeSessionId,
-				method,
-				payload,
-			});
-		};
-
-		const dialogDefault = <T>(opts: ExtensionUIDialogOptions | undefined, fallback: T): Promise<T> => {
-			if (opts?.signal?.aborted) {
-				return Promise.resolve(fallback);
-			}
-			return new Promise((resolveDialog) => {
-				let timeoutId: ReturnType<typeof setTimeout> | undefined;
-				const cleanup = () => {
-					if (timeoutId) {
-						clearTimeout(timeoutId);
-					}
-					opts?.signal?.removeEventListener("abort", onAbort);
-				};
-				const finish = () => {
-					cleanup();
-					resolveDialog(fallback);
-				};
-				const onAbort = () => finish();
-				opts?.signal?.addEventListener("abort", onAbort, { once: true });
-				timeoutId = setTimeout(finish, opts?.timeout ?? 0);
-			});
-		};
-
-		return {
-			select: (title, values, opts) => {
-				emitUiRequest("select", { title, options: values, timeout: opts?.timeout });
-				return dialogDefault(opts, undefined);
-			},
-			confirm: (title, message, opts) => {
-				emitUiRequest("confirm", { title, message, timeout: opts?.timeout });
-				return dialogDefault(opts, false);
-			},
-			input: (title, placeholder, opts) => {
-				emitUiRequest("input", { title, placeholder, timeout: opts?.timeout });
-				return dialogDefault(opts, undefined);
-			},
-			notify: (message, notifyType) => emitUiRequest("notify", { message, notifyType }),
-			onTerminalInput: () => () => {},
-			setStatus: (key, text) => emitUiRequest("setStatus", { statusKey: key, statusText: text }),
-			setWorkingMessage: (message) => emitUiRequest("setWorkingMessage", { message }),
-			setWorkingVisible: (visible) => emitUiRequest("setWorkingVisible", { visible }),
-			setWorkingIndicator: (indicatorOptions?: WorkingIndicatorOptions) =>
-				emitUiRequest("setWorkingIndicator", { options: indicatorOptions }),
-			setHiddenThinkingLabel: (label) => emitUiRequest("setHiddenThinkingLabel", { label }),
-			setWidget: (key: string, content: unknown, widgetOptions?: ExtensionWidgetOptions) => {
-				if (content === undefined || Array.isArray(content)) {
-					emitUiRequest("setWidget", {
-						widgetKey: key,
-						widgetLines: content,
-						widgetPlacement: widgetOptions?.placement,
-					});
-				}
-			},
-			setFooter: () => {},
-			setHeader: () => {},
-			setTitle: (title) => emitUiRequest("setTitle", { title }),
-			async custom<T>(): Promise<T> {
-				return undefined as T;
-			},
-			pasteToEditor: (text) => emitUiRequest("setEditorText", { text }),
-			setEditorText: (text) => emitUiRequest("setEditorText", { text }),
-			getEditorText: () => "",
-			editor: (title, prefill) => {
-				emitUiRequest("editor", { title, prefill });
-				return Promise.resolve(undefined);
-			},
-			addAutocompleteProvider: () => {},
-			setEditorComponent: () => {},
-			getEditorComponent: () => undefined,
-			get theme(): Theme {
-				return theme;
-			},
-			getAllThemes: () => [],
-			getTheme: () => undefined,
-			setTheme: () => ({ success: false, error: "Theme switching is not supported in daemon mode" }),
-			getToolsExpanded: () => false,
-			setToolsExpanded: () => {},
-		};
-	}
-
 	private getRecord(id: string): DaemonSessionRecord {
 		const direct = this.sessions.get(id);
 		if (direct) {
@@ -489,7 +155,7 @@ class AgentDaemon {
 	}
 
 	private handleConnection(socket: Socket): void {
-		const client: DaemonClient = {
+		const client: DaemonSocketClient = {
 			id: randomUUID().slice(0, 8),
 			socket,
 			attachedActiveSessionIds: new Set(),
@@ -511,7 +177,7 @@ class AgentDaemon {
 		socket.on("error", cleanup);
 	}
 
-	private async handleLine(client: DaemonClient, line: string): Promise<void> {
+	private async handleLine(client: DaemonSocketClient, line: string): Promise<void> {
 		let command: DaemonCommand;
 		try {
 			command = JSON.parse(line) as DaemonCommand;
@@ -530,7 +196,10 @@ class AgentDaemon {
 		}
 	}
 
-	private async handleCommand(client: DaemonClient, command: DaemonCommand): Promise<DaemonResponse | undefined> {
+	private async handleCommand(
+		client: DaemonSocketClient,
+		command: DaemonCommand,
+	): Promise<DaemonResponse | undefined> {
 		switch (command.type) {
 			case "list":
 				return success(command.id, "list", { sessions: Array.from(this.sessions.values()).map(summaryForRecord) });
@@ -679,13 +348,13 @@ class AgentDaemon {
 		}
 	}
 
-	private detachClientFromRecord(client: DaemonClient, record: DaemonSessionRecord): void {
+	private detachClientFromRecord(client: DaemonSocketClient, record: DaemonSessionRecord): void {
 		record.clients.delete(client);
 		client.attachedActiveSessionIds.delete(record.activeSessionId);
 		this.write(client, { type: "session_detached", activeSessionId: record.activeSessionId });
 	}
 
-	private detachClient(client: DaemonClient): void {
+	private detachClient(client: DaemonSocketClient): void {
 		for (const activeSessionId of [...client.attachedActiveSessionIds]) {
 			const record = this.sessions.get(activeSessionId);
 			if (record) {
@@ -711,7 +380,7 @@ class AgentDaemon {
 		}
 	}
 
-	private write(client: DaemonClient, message: DaemonOutbound): void {
+	private write(client: DaemonSocketClient, message: DaemonOutbound): void {
 		if (client.socket.destroyed) {
 			return;
 		}
