+++ packages/coding-agent/src/modes/index.ts
@@ -3,16 +3,16 @@
  */
 
 export { DaemonClient, type DaemonClientMessageListener } from "./daemon/daemon-client.js";
-export {
-	type DaemonCommand,
-	type DaemonModeOptions,
-	type DaemonOutbound,
-	type DaemonResponse,
-	type DaemonSessionState,
-	type DaemonSessionSummary,
-	defaultDaemonSocketPath,
-	runDaemonMode,
-} from "./daemon/daemon-mode.js";
+export { runDaemonMode } from "./daemon/daemon-mode.js";
+export type {
+	DaemonCommand,
+	DaemonModeOptions,
+	DaemonOutbound,
+	DaemonResponse,
+	DaemonSessionState,
+	DaemonSessionSummary,
+} from "./daemon/daemon-protocol.js";
+export { defaultDaemonSocketPath } from "./daemon/daemon-socket.js";
 export { InteractiveMode, type InteractiveModeOptions } from "./interactive/interactive-mode.js";
 export { type PrintModeOptions, runPrintMode } from "./print-mode.js";
 export { type ModelInfo, RpcClient, type RpcClientOptions, type RpcEventListener } from "./rpc/rpc-client.js";
